RelativeAgeEstimationAnonymous/Relative Age Estimation/requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
black==22.1.0
bleach==4.1.0
cachetools==4.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
condor-pytorch==1.1.0
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
einops==0.3.2
entrypoints==0.3
executing==0.8.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
h5py==3.6.0
idna==2.10
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.1
llvmlite==0.35.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.3
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numba==0.52.0
numpy==1.21.5
oauthlib==3.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.0.1
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.14.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2021.3
#pywin32==303
#pywinpty==2.0.1
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
stack-data==0.1.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==2.1.0
tomli==2.0.0
#torch==1.10.1+cu113
#torchaudio==0.10.1+cu113
#torchvision==0.11.2+cu113
#torch==1.13.1 
#torchvision==0.14.1 
#torchaudio==0.13.1
tornado==6.1
tqdm==4.55.0
traitlets==5.1.1
typing_extensions==4.0.1
umap-learn==0.4.6
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.7.0